# docs/swagger.yaml
swagger: '2.0'
info:
  title: flow API
  description: flow version - not yet reached
  version: 0.1.0
host: localhost:8881
schemes:
  - http
basePath: /
produces:
  - application/json
paths:
  /users/login:
    get:
      summary: Get a user's aid
      parameters:
        - name: username
          in: query
          description: The username of the user
          required: true
          type: string
        - name: password
          in: query
          description: The password of the user
          required: true
          type: string
      tags:
        - Users
      responses:
        200:
          description: A json object
          schema:
            type: object
            properties:
              aid:
                type: string
        400:
          $ref: '#/responses/400'
    post:
      summary: Create a user account & get the aid
      consumes:
        - application/json
      parameters:
        - name: body
          in: body
          required: true
          schema:
            properties:
              username:
                type: string
              password:
                type: string
      tags:
        - Users
      responses:
        200:
          description: A json object
          schema:
            type: object
            properties:
              aid:
                type: string
        400:
          $ref: '#/responses/400'
          
  /users/join:
    get:
      summary: Create a user account & get the aid
      parameters:
        - name: username
          in: query
          description: The username of the user
          required: true
          type: string
        - name: password
          in: query
          description: The password of the user
          required: true
          type: string
      tags:
        - Users
      responses:
        200:
          description: A json object
          schema:
            type: object
            properties:
              aid:
                type: string
        400:
          $ref: '#/responses/400'
    post:
      summary: Create a user account & get the aid
      consumes:
        - application/json
      parameters:
        - name: body
          in: body
          required: true
          schema:
            properties:
              username:
                type: string
              password:
                type: string
      tags:
        - Users
      responses:
        200:
          description: A json object
          schema:
            type: object
            properties:
              aid:
                type: string
        400:
          $ref: '#/responses/400'

  /users/metrics:
    get:
      summary: Get the latest metrics on users
      responses:
        200:
          description: a json object
          schema:
            type: object
            properties:
              RegisteredUsers:
                type: number

responses:
  400:
    description: Error
    schema:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        fields:
          type: string